test: place credit pool migration in testcontainers
Copy the finalized migrated credit pool tests from the old unit-test path into the Testcontainers integration-test path, then remove the old file.

Review command:

  git diff --quiet HEAD^:api/tests/unit_tests/services/test_credit_pool_service.py HEAD:api/tests/test_containers_integration_tests/services/test_credit_pool_service.py

That command should exit 0, proving the placed Testcontainers file is byte-identical to the reviewed in-place migration from the previous commit.
